+++ vector/src/main/java/im/vector/riotx/core/extensions/Session.kt
@@ -16,30 +16,50 @@
 
 package im.vector.riotx.core.extensions
 
+import android.content.Context
+import androidx.core.content.ContextCompat
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.ProcessLifecycleOwner
 import im.vector.matrix.android.api.session.Session
 import im.vector.matrix.android.api.session.crypto.keysbackup.KeysBackupState
 import im.vector.matrix.android.api.session.sync.FilterService
+import im.vector.riotx.core.services.VectorSyncService
 import im.vector.riotx.features.notifications.PushRuleTriggerListener
 import im.vector.riotx.features.session.SessionListener
 import timber.log.Timber
 
-fun Session.configureAndStart(pushRuleTriggerListener: PushRuleTriggerListener,
+fun Session.configureAndStart(context: Context,
+                              pushRuleTriggerListener: PushRuleTriggerListener,
                               sessionListener: SessionListener) {
     open()
     addListener(sessionListener)
     setFilter(FilterService.FilterPreset.RiotFilter)
     Timber.i("Configure and start session for ${this.myUserId}")
-    val isAtLeastStarted = ProcessLifecycleOwner.get().lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)
-    Timber.v("--> is at least started? $isAtLeastStarted")
-    startSync(isAtLeastStarted)
+    startSyncing(context)
     refreshPushers()
     pushRuleTriggerListener.startWithSession(this)
 
     // TODO P1 From HomeActivity
     // @Inject lateinit var incomingVerificationRequestHandler: IncomingVerificationRequestHandler
     // @Inject lateinit var keyRequestHandler: KeyRequestHandler
+}
+
+fun Session.startSyncing(context: Context) {
+    val applicationContext = context.applicationContext
+    if (!hasAlreadySynced()) {
+        VectorSyncService.newIntent(applicationContext, myUserId).also {
+            try {
+                ContextCompat.startForegroundService(applicationContext, it)
+            } catch (ex: Throwable) {
+                // TODO
+                Timber.e(ex)
+            }
+        }
+    } else {
+        val isAtLeastStarted = ProcessLifecycleOwner.get().lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)
+        Timber.v("--> is at least started? $isAtLeastStarted")
+        startSync(isAtLeastStarted)
+    }
 }
 
 /**