+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/timeline/TimelineEventController.kt
@@ -220,7 +220,7 @@ class TimelineEventController @Inject constructor(private val dateFormatter: Vec
         showingForwardLoader = LoadingItem_()
                 .id("forward_loading_item_$timestamp")
                 .setVisibilityStateChangedListener(Timeline.Direction.FORWARDS)
-                .addWhen(Timeline.Direction.FORWARDS)
+                .addWhenLoading(Timeline.Direction.FORWARDS)
 
         val timelineModels = getModels()
         add(timelineModels)
@@ -230,14 +230,18 @@ class TimelineEventController @Inject constructor(private val dateFormatter: Vec
             LoadingItem_()
                     .id("backward_loading_item_$timestamp")
                     .setVisibilityStateChangedListener(Timeline.Direction.BACKWARDS)
-                    .addWhen(Timeline.Direction.BACKWARDS)
+                    .addWhenLoading(Timeline.Direction.BACKWARDS)
         }
     }
 
 // Timeline.LISTENER ***************************************************************************
 
-    override fun onUpdated(snapshot: List<TimelineEvent>) {
+    override fun onTimelineUpdated(snapshot: List<TimelineEvent>) {
         submitSnapshot(snapshot)
+    }
+
+    override fun onTimelineFailure(throwable: Throwable) {
+        // no-op, already handled
     }
 
     private fun submitSnapshot(newSnapshot: List<TimelineEvent>) {
@@ -247,6 +251,7 @@ class TimelineEventController @Inject constructor(private val dateFormatter: Vec
             currentSnapshot = newSnapshot
             val diffResult = DiffUtil.calculateDiff(diffCallback)
             diffResult.dispatchUpdatesTo(listUpdateCallback)
+            requestDelayedModelBuild(100)
             inSubmitList = false
         }
     }
@@ -319,7 +324,7 @@ class TimelineEventController @Inject constructor(private val dateFormatter: Vec
     /**
      * Return true if added
      */
-    private fun LoadingItem_.addWhen(direction: Timeline.Direction): Boolean {
+    private fun LoadingItem_.addWhenLoading(direction: Timeline.Direction): Boolean {
         val shouldAdd = timeline?.hasMoreToLoad(direction) ?: false
         addIf(shouldAdd, this@TimelineEventController)
         return shouldAdd